content_demo/requirements.txt
alembic==1.10.2
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
fastapi==0.95.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
isort==5.12.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
psycopg2==2.9.5
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
python-dotenv==1.0.0
redis==4.5.4
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.8
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
